improve and style drawer
